Refactor map code to be reusable and composable
