+++ app/templates/app/map_page.html
@@ -1,123 +1,11 @@
 {% extends "base.html" %}
 
-{% load menu_tags humanize brand groundwork_geo wagtailsettings_tags account setting django_bootstrap5 wagtailcore_tags static wagtailroutablepage_tags django_bootstrap5 setting %}
-
 {% block content_padding %}p-0{% endblock %}
 
 {% block content_width %}container-fluid{% endblock %}
 
 {% block content %}
-  {% map in_place=False center="[-2.5, 53.6]" zoom=5.5 style="mapbox://styles/commonknowledge/cl7cnn4d6004a14nzcmvf0k01/draft" %}
-  <main
-    class='tw-flex tw-flex-col md:tw-grid md:tw-grid-cols-4 xl:tw-grid-cols-5'
-    data-map-target="config"
-    data-controller="zoom-to-source-features"
-    data-zoom-to-source-features-source-ids-value='["events"]'
-    data-zoom-to-source-features-max-zoom-value="12"
-  >
-    <section
-      class='tw-col-span-2 tw-order-last md:tw-order-first'
-      data-controller="list-filter map-html-bridge"
-      data-map-html-bridge-html-scroll-query-value="html"
-      data-map-html-bridge-map-layer-value="event-icons"
-      data-map-html-bridge-map-event-value="mouseenter"
-      data-map-html-bridge-html-id-prefix-value="event-"
-      data-map-html-bridge-map-layer-id-property-value="id"
-      data-map-target="config"
-      data-list-filter-selected-value='{ "locationType": "__ALL__" }'
-      data-list-filter-option-active-classes-value="tw-bg-yellow"
-      data-list-filter-option-passive-classes-value="tw-bg-slate-300"
-    >
-      <article class="px-4 my-3">
-        <h1 class='tw-text-4xl'>{{ page.title }}</h1>
-        <div class='tw-text-lg'>
-          {{ page.intro|richtext }}
-        </div>
-        {% if not user.is_member %}
-          {% bootstrap_button href="/join" content="Join LBC and come along" %}
-        {% endif %}
-      </article>
-      <div class="tw-space-x-1 px-4 tw-text-sm my-3">
-        <a
-          data-list-filter-target="option"
-          data-list-filter-value-param="__ALL__"
-          data-list-filter-attr-param="locationType"
-          data-action="click->list-filter#selectForAttr" 
-          class="tw-cursor-pointer tw-no-underline tw-text-black tw-inline-block tw-py-1 tw-px-2 tw-rounded-lg tw-bg-slate-300 hover:tw-bg-slate-400 hover:tw-text-black">All</a>
-        <a
-          data-list-filter-target="option"
-          data-list-filter-value-param="in_person"
-          data-list-filter-attr-param="locationType"
-          data-action="click->list-filter#selectForAttr" 
-          class="tw-cursor-pointer tw-no-underline tw-text-black tw-inline-block tw-py-1 tw-px-2 tw-rounded-lg tw-bg-slate-300 hover:tw-bg-slate-400 hover:tw-text-black">In Person</a>
-        <a
-          data-list-filter-target="option"
-          data-list-filter-value-param="virtual"
-          data-list-filter-attr-param="locationType"
-          data-action="click->list-filter#selectForAttr" 
-          class="tw-cursor-pointer tw-no-underline tw-text-black tw-inline-block tw-py-1 tw-px-2 tw-rounded-lg tw-bg-slate-300 hover:tw-bg-slate-400 hover:tw-text-black">Online</a>
-        <span class='tw-rounded-lg tw-bg-slate-100'>
-          <span data-list-filter-target="count">0</span> events
-        </span>
-      </div>
-      <ol class="tw-relative tw-list-none tw-p-0 tw-m-0 tw-space-y-2 tw-mb-2">
-      {% for event in events %}
-        <li
-          data-map-html-bridge-zoom="12"
-          data-map-html-bridge-lng-param="{{ event.geometry.coordinates.0 }}"
-          data-map-html-bridge-lat-param="{{ event.geometry.coordinates.1 }}"
-          data-action="mouseenter->map-html-bridge#flyTo"
-          id="event-{{event.properties.id}}"
-          data-list-filter-target="item"
-          data-location-type="{{ event.properties.location_type }}"
-          class="tw-px-2">
-            <div class='tw-flex tw-flex-row tw-bg-white'>
-              <div class="tw-w-[60px] tw-text-center tw-relative tw-flex-shrink-0">
-                <div class='tw-sticky tw-top-0 tw-py-3 tw-whitespace-nowrap tw-overflow-hidden tw-leading-5'>
-                  <div class='tw-text-lg'>{{ event.properties.starts_at|date:"j" }}</div>
-                  <div class='tw-uppercase '>{{ event.properties.starts_at|date:"M" }}</div>
-                </div>
-              </div>
-              <div class="tw-p-3 tw-pl-1 tw-mb-1 tw-overflow-hidden tw-break-words tw-border-r-2 tw-border-r-transparent hover:tw-border-r-yellow tw-transition-all">
-                <h3 class="tw-text-2xl tw-font-semibold tw-text-gray-900">{{ event.properties.name }}</h3>
-                <div class='tw-text-sm tw-font-normal tw-leading-none tw-my-3 tw-text-gray-700'>
-                  <div class="tw-mb-2">
-                    <span class="tw-capitalize">{{ event.properties.location_type|replace:"_| " }}</span>
-                    <span>&middot;</span>
-                    <span>Starts at {{event.properties.starts_at|date:"P"}}</span>
-                    <span>&middot;</span>
-                    <span>{{event.properties.starts_at|naturaltime}}</span>
-                  </div>
-                  {% if event.properties.physical_address.formatted_address %}
-                  <div>
-                    {{ event.properties.physical_address.formatted_address }}
-                  </div>
-                  {% endif %}
-                </div>
-                {% if event.properties.body.body %}
-                  {{ event.properties.body.body|safe }}
-                {% endif %}
-                <br />
-                {% bootstrap_button href=event.properties.url content="Learn more" %}
-              </div>
-            </div>
-        </li>
-      {% endfor %}
-      </ol>
-    </section>
-    <section class='tw-h-[250px] tw-relative md:tw-sticky md:tw-top-0 md:tw-h-screen md:tw-col-span-2 xl:tw-col-span-3'>
-        {% for id, data in sources.items %}
-          {% map_source id=id data=data %}
-        {% endfor %}
-      
-        {% for id, layer in layers.items %}
-          {% map_layer id=id layer=layer %}
-        {% endfor %}
-
-        <div data-controller="scroll-into-view-on-event" data-map-target="canvas" class='tw-h-full tw-w-full tw-absolute tw-top-0 tw-left-0'></div>
-    </section>
-  </main>
-  {% endmap %}
+  {% include "app/includes/event_list_and_map.html" with title=page.title intro=page.intro %}
 {% endblock %}
 
 {% block footer %}{% endblock %}
